backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
supabase==2.0.2
redis==5.0.1

# HTTP Client
httpx==0.24.1

# Authentication
python-jose[cryptography]==3.3.0

# AI Integration
groq==0.4.1

# WebSocket support
websockets==12.0

# Audio (optional)
elevenlabs==0.2.26

# Utilities
email-validator>=1.1.0